==> analysis/04_diagnosis_qc_gigamuga_12_batches.Rmd <==
---
title: "diagnosis_qc_gigamuga_12_batches"
author: "Hao He"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Genotype diagnostics for diversity outbred mice
We first load the R/qtl2 package and the data.
We’ll also load the R/broman package for some utilities and plotting functions, and R/qtlcharts for interactive graphs.

# library
```{r library, message=FALSE}
library(broman)
library(qtl2)
library(qtlcharts)
library(ggplot2)
library(ggrepel)
library(DOQTL)
library(mclust)
library(tidyverse)
library(reshape2)
library(DT)
source("code/reconst_utils.R")
options(stringsAsFactors = F)
```

# Generate json file for all 12 batches
```{r generate json file for all 12 batches, eval=TRUE}
#total sample id
#load json file for the 12 batches
gm <- get(load("data/Jackson_Lab_12_batches/gm_12batches.RData"))

gm
```

# Missing data per sample
```{r Missing data per sample,fig.height=12, fig.width=20, eval=TRUE}
percent_missing <- n_missing(gm, "ind", "prop")*100
miss_dat <- data.frame(Mouse=seq_along(percent_missing),
                       id = names(percent_missing),
                       Percent_missing_genotype_data = percent_missing,
                       batch = as.character(do.call(rbind.data.frame, 
                                                      strsplit(ind_ids(gm), "_"))[,5]),
                       labels = as.character(do.call(rbind.data.frame, 
                                                       strsplit(ind_ids(gm), "V01_"))[,2]))
miss_dat <- miss_dat %>%
  mutate(labels2 = case_when(
    percent_missing <= 10 ~ "",
    TRUE ~ labels
  ))

#iplot
iplot(miss_dat$Mouse, 
      miss_dat$Percent_missing_genotype_data, 
      indID=paste0(miss_dat$labels, " (", round(miss_dat$Percent_missing_genotype_data,2), "%)"),
      chartOpts=list(xlab="Mouse", 
                     ylab="Percent missing genotype data",
                     ylim=c(0, 100)))

#save into pdf
pdf(file = "data/Jackson_Lab_12_batches/Percent_missing_genotype_data.pdf", width = 20, height = 20)
# Change point shapes and colors
p <- ggplot(data = miss_dat, 
        aes(x=Mouse, y=Percent_missing_genotype_data, color = batch)) +
  geom_point() +
  geom_hline(yintercept=5, linetype="solid", color = "red") +
  geom_text_repel(aes(label=labels2), vjust = 0, nudge_y = 0.01, show.legend = FALSE, size=3) +
  theme(text = element_text(size = 20))
p
dev.off()

p

save(percent_missing,
     file = "data/Jackson_Lab_12_batches/percent_missing_id.RData")
```

# Sexes
```{r Sexes, eval=TRUE}
xint <- read_csv_numer("data/Jackson_Lab_12_batches/Jackson_Lab_12_batches_qtl2_chrXint.csv", transpose=TRUE)
yint <- read_csv_numer("data/Jackson_Lab_12_batches/Jackson_Lab_12_batches_qtl2_chrYint.csv", transpose=TRUE)

# Gigamuga marker annotation file from UNC.
gm_marker_file = "http://csbio.unc.edu/MUGA/snps.gigamuga.Rdata" #FIXED
# Read in the UNC GigaMUGA SNPs and clusters.
load(url(gm_marker_file))
#subset down to gm
snps$marker = as.character(snps$marker)

#load the intensities.fst.RData
load("data/Jackson_Lab_12_batches/intensities.fst.RData")
#X and Y channel
X <- result[result$channel == "X",]
rownames(X) <- X$snp
X <- X[,c(-1,-2)]

Y <- result[result$channel == "Y",]
rownames(Y) <- Y$snp
Y <- Y[,c(-1,-2)]

#determine predict.sex
predict.sex = determine_sex_chry_m(x = X, y = Y, markers = snps)$sex

gm$covar <- gm$covar %>%
  mutate(id = rownames(gm$covar)) %>%
  left_join(data.frame(id = names(predict.sex),
                       predict.sex = predict.sex,stringsAsFactors = F))
rownames(gm$covar) <- gm$covar$id

#sex order
sex <- gm$covar[rownames(xint),"sex"]

x_pval <- apply(xint, 2, function(a) t.test(a ~ sex)$p.value)
y_pval <- apply(yint, 2, function(a) t.test(a ~ sex)$p.value)

xint_ave <- rowMeans(xint[, x_pval < 0.05/length(x_pval)], na.rm=TRUE)
yint_ave <- rowMeans(yint[, y_pval < 0.05/length(y_pval)], na.rm=TRUE)

point_colors <- as.character( brocolors("web")[c("green", "purple")] )
labels <- paste0(names(xint_ave))
iplot(xint_ave, yint_ave, group=sex, indID=labels,
      chartOpts=list(pointcolor=point_colors, pointsize=4,
                     xlab="Average X chr intensity", ylab="Average Y chr intensity"))

phetX <- rowSums(gm$geno$X == 2)/rowSums(gm$geno$X != 0)
iplot(xint_ave, phetX, group=sex, indID=labels,
      chartOpts=list(pointcolor=point_colors, pointsize=4,
                     xlab="Average X chr intensity", ylab="Proportion het on X chr"))
```

# Sample duplicates
```{r Sample duplicates, eval=TRUE, fig.height=12, fig.width=16}
cg <- compare_geno(gm, cores=10)
summary.cg <- summary(cg, threshold = 0)
#get the name and missing percentage
summary.cg$Name.ind1 <- str_split_fixed(summary.cg$ind1, "_",7)[,6]
summary.cg$Name.ind2 <- str_split_fixed(summary.cg$ind2, "_",7)[,6]
summary.cg$miss.ind1 <- percent_missing[match(summary.cg$ind1, names(percent_missing))]
summary.cg$miss.ind2 <- percent_missing[match(summary.cg$ind2, names(percent_missing))]
summary.cg$remove.id <- ifelse(summary.cg$miss.ind1 > summary.cg$miss.ind2, summary.cg$ind1, summary.cg$ind2)

#filter prop_match>=0.85 or same name for Name.ind1 and Name.ind2
filtered.summary.cg <- summary.cg %>%
  mutate(same.sample = case_when(
    Name.ind1 == Name.ind2 ~ TRUE,
    Name.ind1 != Name.ind2 ~ FALSE
  )) %>%
  filter(prop_match >= 0.85 | same.sample == TRUE)

save(filtered.summary.cg,
     file = "data/Jackson_Lab_12_batches/filtered.summary.cg.RData")

#display filtered.summary.cg
DT::datatable(filtered.summary.cg, filter = list(position = 'top', clear = FALSE),
              options = list(pageLength = 40, scrollY = "300px", scrollX = "40px"))

#plot prop matrix for same.sample = false and prop_match >= 0.85
filter.id <- data.frame(id = unique(c(filtered.summary.cg[filtered.summary.cg$same.sample == F,]$ind1, 
                                      filtered.summary.cg[filtered.summary.cg$same.sample == F,]$ind2)))
filter.id$name <- do.call(rbind.data.frame, 
                              strsplit(filter.id$id, "V01_"))[,2]
filter.id <- filter.id[order(filter.id$name),]
gm_filter <- gm[filter.id$id,]
#replace id names
old_ids <- do.call(rbind.data.frame, 
                              strsplit(ind_ids(gm_filter), "V01_"))[,2]
new_ids <- setNames(old_ids,
                    ind_ids(gm_filter))
gm_filter <- replace_ids(gm_filter, new_ids)
#save gm_filter for  same.sample = false and prop_match >= 0.85
save(gm_filter,  file = "data/Jackson_Lab_12_batches/gm_filterprop_match_0.85.RData")

#compare geno
filter.cg <- compare_geno(gm_filter, cores=10, proportion = TRUE)
filter.cg[lower.tri(filter.cg)] <- NA
filter.cg[filter.cg < 0.5] <- NA # for ggplot lowest value 0.5
diag(filter.cg) <- 0
# Melt the correlation matrix
melted_cormat <- melt(filter.cg, na.rm = TRUE)
# Heatmap
p <- ggplot(data = melted_cormat, aes(Var2, Var1, fill = value))+
 geom_tile(color = "white")+
 scale_fill_gradient2(low = "white", high = "red", 
   limit = c(0.5,1), space = "Lab", 
   name="Proportions of matching genotypes") +
 scale_y_discrete(position = "right") +
 xlab("") +
 ylab("") +
 theme_bw() + 
 theme(panel.border = element_blank(), panel.grid.major = element_blank(),
       panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +
 theme(axis.text.x = element_text(angle = 90, vjust = 0.5, 
                                  size = 10, hjust = 1),
       axis.text.y = element_text(size = 10)) +
 coord_fixed()
p

pdf(file = "data/Jackson_Lab_12_batches/Proportion_matching_genotypes_before_removal_samples.pdf", width = 20, height = 20) 
par(mar=c(5.1,0.6,0.6, 0.6))
hist(cg[upper.tri(cg)], breaks=seq(0, 1, length=201),
     main="", yaxt="n", ylab="", xlab="Proportion matching genotypes")
rug(cg[upper.tri(cg)])
dev.off()

par(mar=c(5.1,0.6,0.6, 0.6))
hist(cg[upper.tri(cg)], breaks=seq(0, 1, length=201),
     main="", yaxt="n", ylab="", xlab="Proportion matching genotypes")
rug(cg[upper.tri(cg)])

pdf(file = "data/Jackson_Lab_12_batches/Proportion_matching_genotypes_after_removal_samples_percent_missing_5.pdf",width = 20, height = 20) 
cgsub <- cg[percent_missing < 5, percent_missing < 5]
par(mar=c(5.1,0.6,0.6, 0.6))
hist(cgsub[upper.tri(cgsub)], breaks=seq(0, 1, length=201),
     main="", yaxt="n", ylab="", xlab="Proportion matching genotypes")
rug(cgsub[upper.tri(cgsub)])
dev.off()

cgsub <- cg[percent_missing < 5, percent_missing < 5]
par(mar=c(5.1,0.6,0.6, 0.6))
hist(cgsub[upper.tri(cgsub)], breaks=seq(0, 1, length=201),
     main="", yaxt="n", ylab="", xlab="Proportion matching genotypes")
rug(cgsub[upper.tri(cgsub)])

#show samples with missing genotypes >5
miss_dat_5 <- miss_dat %>% 
  arrange(desc(Percent_missing_genotype_data)) %>%
  filter(labels2 != "")
dim(miss_dat_5)
#display miss_dat
DT::datatable(miss_dat_5,filter = list(position = 'top', clear = FALSE),
              options = list(pageLength = 40, scrollY = "300px", scrollX = "40px"))
```

# Array intensities and Genotype frequencies
```{r Array intensities and Genotype frequencies, fig.height=12, fig.width=12, eval=TRUE}
#result object is Array intensities 286518*3284
result <- result[seq(1, nrow(result), by=2),-(1:2)] + result[-seq(1, nrow(result), by=2),-(1:2)]
result <- result[,intersect(ind_ids(gm), colnames(result))]
n <- names(sort(percent_missing[intersect(ind_ids(gm), colnames(result))], decreasing=TRUE))
iboxplot(log10(t(result[,n])+1), orderByMedian=FALSE, chartOpts=list(ylab="log10(SNP intensity + 1)"))

# Genotype frequencies
g <- do.call("cbind", gm$geno[1:19])
fg <- do.call("cbind", gm$founder_geno[1:19])
g <- g[,colSums(fg==0)==0]
fg <- fg[,colSums(fg==0)==0]
fgn <- colSums(fg==3)

gf_ind <- vector("list", 4)
for(i in 1:4) {
  gf_ind[[i]] <- t(apply(g[,fgn==i], 1, function(a) table(factor(a, 1:3))/sum(a != 0)))
}

par(mfrow=c(4,1), mar=c(0.6, 0.6, 2.6, 0.6))
for(i in 1:4) {
  triplot(c("AA", "AB", "BB"), main=paste0("MAF = ", i, "/8"))
  tripoints(gf_ind[[i]], pch=21, bg="lightblue")
  tripoints(c((1-i/8)^2, 2*i/8*(1-i/8), (i/8)^2), pch=21, bg="violetred")
  
  if(i>=3) { # label mouse with lowest het
    wh <- which(gf_ind[[i]][,2] == min(gf_ind[[i]][,2]))
    tritext(gf_ind[[i]][wh,,drop=FALSE] + c(0.02, -0.02, 0),
            names(wh), adj=c(0, 1))
  }
  
  # label other mice
  if(i==1) {
    lab <- rownames(gf_ind[[i]])[gf_ind[[i]][,2]>0.3]
  }
  else if(i==2) {
    lab <- rownames(gf_ind[[i]])[gf_ind[[i]][,2]>0.48]
  }
  else if(i==3) {
    lab <- rownames(gf_ind[[i]])[gf_ind[[i]][,2]>0.51]
  }
  else if(i==4) {
    lab <- rownames(gf_ind[[i]])[gf_ind[[i]][,2]>0.6]
  }
  
  for(ind in lab) {
    if(grepl("^F", ind) && i != 3) {
      tritext(gf_ind[[i]][ind,,drop=FALSE] + c(-0.01, 0, +0.01), ind, adj=c(1,0.5))
    } else {
      tritext(gf_ind[[i]][ind,,drop=FALSE] + c(0.01, 0, -0.01), ind, adj=c(0,0.5))
    }
  }
}

```

# Crossover counts and Genotyping error LOD scores
```{r Crossover counts and Genotyping error LOD scores, eval=TRUE, fig.height=12, fig.width=14}
#load pre-caluated results
load("data/Jackson_Lab_12_batches/pr.RData")
load("data/Jackson_Lab_12_batches/m.RData")
load("data/Jackson_Lab_12_batches/nxo.RData")

#crossover
totxo <- rowSums(nxo)[ind_ids(gm)]
all.equal(ind_ids(gm), names(totxo))
iplot(seq_along(totxo),
      totxo,
      group=gm$covar$ngen,
      chartOpts=list(xlab="Mouse", ylab="Number of crossovers", 
                     margin=list(left=80,top=40,right=40,bottom=40,inner=5),
                     axispos=list(xtitle=25,ytitle=50,xlabel=5,ylabel=5)))

#save crossover into pdf
pdf(file = "data/Jackson_Lab_12_batches/number_crossover.pdf")
cross_over <- data.frame(Mouse = seq_along(totxo), Number_crossovers = totxo, generation = gm$covar$ngen)
names(totxo) <- as.character(do.call(rbind.data.frame, strsplit(names(totxo), "V01_"))[,2])
names(totxo)[totxo <= 800 & totxo >= 400] = ""
# Change point shapes and colors
p <-ggplot(cross_over, aes(x=Mouse, y=Number_crossovers, fill = generation, color=generation)) +
  geom_point() +
  geom_text_repel(aes(label=names(totxo),hjust=0,vjust=0), show.legend = FALSE)
p
dev.off()

p

#Here are the crossover counts for those  mice with percent_missing >= 5:
tmp <- cbind(percent_missing=round(percent_missing,2), total_xo=totxo)[percent_missing >= 5,]
#display miss_dat
DT::datatable(tmp[order(tmp[,1]),], filter = list(position = 'top', clear = FALSE),
              options = list(pageLength = 40, scrollY = "300px", scrollX = "40px"))

# Genotyping error LOD scores
load("data/Jackson_Lab_12_batches/e.RData")
errors_ind <- rowSums(e>2)/n_typed(gm)*100
lab <- paste0(names(errors_ind), " (", myround(percent_missing,1), "%)")
iplot(seq_along(errors_ind), errors_ind, indID=lab,
      chartOpts=list(xlab="Mouse", ylab="Percent genotyping errors", ylim=c(0, 8),
                     axispos=list(xtitle=25, ytitle=50, xlabel=5, ylabel=5)))
save(errors_ind, file = "data/Jackson_Lab_12_batches/errors_ind.RData")

# Apparent genotyping errors
load("data/Jackson_Lab_12_batches/snpg.RData")

gobs <- do.call("cbind", gm$geno)
gobs[gobs==0] <- NA

par(pty="s")
err_direct <- rowMeans(snpg != gobs, na.rm=TRUE)*100
errors_ind_0 <- rowSums(e > 0)/n_typed(gm)*100
par(mar=c(4.1,4.1,0.6, 0.6))
grayplot(errors_ind_0, err_direct,
         xlab="Percent errors (error LOD > 0)",
         ylab="Percent errors (obs vs predicted)",
         xlim=c(0, 2), ylim=c(0, 2))
abline(0,1,lty=2, col="gray60")

pdf(file = "data/Jackson_Lab_12_batches/Percent_genotype_errors_obs_vs_predicted.pdf",width = 20, height = 20) 
par(pty="s")
err_direct <- rowMeans(snpg != gobs, na.rm=TRUE)*100
errors_ind_0 <- rowSums(e > 0)/n_typed(gm)*100
par(mar=c(4.1,4.1,0.6, 0.6))
grayplot(errors_ind_0, err_direct,
         xlab="Percent errors (error LOD > 0)",
         ylab="Percent errors (obs vs predicted)",
         xlim=c(0, 2), ylim=c(0, 2))
abline(0,1,lty=2, col="gray60")
dev.off()
```

# Missing data in Markers and Genotype frequencies Markers
```{r Missing data in Markers and Genotype frequencies Markers, eval=TRUE}
#It can also be useful to look at the proportion of missing genotypes by marker. 
#Markers with a lot of missing data were likely difficult to call, and so the genotypes that were called may contain a lot of errors.
pmis_mar <- n_missing(gm, "marker", "proportion")*100

par(mar=c(5.1,0.6,0.6, 0.6))
hist(pmis_mar, breaks=seq(0, 100, length=201),
     main="", yaxt="n", ylab="", xlab="Percent missing genotypes")
rug(pmis_mar)

pdf(file = "data/Jackson_Lab_12_batches/Percent_missing_genotype_data_per_marker.pdf")
par(mar=c(5.1,0.6,0.6, 0.6))
hist(pmis_mar, breaks=seq(0, 100, length=201),
     main="", yaxt="n", ylab="", xlab="Percent missing genotypes")
rug(pmis_mar)
dev.off()

# Genotype frequencies Markers
gf_mar <- t(apply(g, 2, function(a) table(factor(a, 1:3))/sum(a != 0)))
gn_mar <- t(apply(g, 2, function(a) table(factor(a, 1:3))))

pdf(file = "data/Jackson_Lab_12_batches/genotype_frequency_marker.pdf")
par(mfrow=c(2,2), mar=c(0.6, 0.6, 2.6, 0.6))
for(i in 1:4) {
  triplot(c("AA", "AB", "BB"), main=paste0("MAF = ", i, "/8"))
  z <- gf_mar[fgn==i,]
  z <- z[rowSums(is.na(z)) < 3,]
  tripoints(z, pch=21, bg="gray80", cex=0.6)
  tripoints(c((1-i/8)^2, 2*i/8*(1-i/8), (i/8)^2), pch=21, bg="violetred")
}
dev.off()

par(mfrow=c(2,2), mar=c(0.6, 0.6, 2.6, 0.6))
for(i in 1:4) {
  triplot(c("AA", "AB", "BB"), main=paste0("MAF = ", i, "/8"))
  z <- gf_mar[fgn==i,]
  z <- z[rowSums(is.na(z)) < 3,]
  tripoints(z, pch=21, bg="gray80", cex=0.6)
  tripoints(c((1-i/8)^2, 2*i/8*(1-i/8), (i/8)^2), pch=21, bg="violetred")
}

# Genotype errors Markers
errors_mar <- colSums(e>2)/n_typed(gm, "marker")*100

grayplot(pmis_mar, errors_mar,
         xlab="Proportion missing", ylab="Proportion genotyping errors")

pdf(file = "data/Jackson_Lab_12_batches/genotype_error_marker.pdf")
grayplot(pmis_mar, errors_mar,
         xlab="Proportion missing", ylab="Proportion genotyping errors")
dev.off()
```

# Remove bad samples
```{r remove bad samples, eval=TRUE}
#qc_infor
#percent missing
qc_info <- left_join(gm$covar, miss_dat)
#add cross_over
cross_over$id <- rownames(cross_over)
qc_info <- qc_info %>% left_join(cross_over[,-1])

#mismatch sex
qc_info <- qc_info %>%
  mutate(sex.match = case_when(
    predict.sex == sex ~ TRUE,
    predict.sex != sex ~ FALSE
  ))

#genotype errors
qc_info <- qc_info %>%
  left_join(
    data.frame(id = names(errors_ind),
               genotype_erros = errors_ind,stringsAsFactors = F)
  )

#add duplicated id to be remove
qc_info <- qc_info %>%
  mutate(remove.id.duplicated = case_when(
    id %in% unique(c(filtered.summary.cg$remove.id)) ~ TRUE,
    !(id %in% unique(c(filtered.summary.cg$remove.id))) ~ FALSE
  ))

#bad sample label
qc_info <- qc_info %>%
    mutate(bad.sample = case_when(
      (ngen ==1 | Number_crossovers <= 200 | Number_crossovers >=1000 | percent_missing >= 10 | genotype_erros >= 1 | remove.id.duplicated == TRUE) ~ TRUE,
      TRUE ~ FALSE
    ))
save(qc_info, file = "data/Jackson_Lab_12_batches/qc_info.RData")

#display qc_info
DT::datatable(qc_info, filter = list(position = 'top', clear = FALSE),
              options = list(pageLength = 40, scrollY = "300px", scrollX = "40px"))

#remove bad samples
gm.no.bad <- gm[paste0("-",as.character(qc_info[qc_info$bad.sample == TRUE, "id"])),]

gm.no.bad

# subjects
# update other stuff
e <- e[ind_ids(gm.no.bad),]
g <- g[ind_ids(gm.no.bad),]
snpg <- snpg[ind_ids(gm.no.bad),]

length(errors_mar[errors_mar > 5])
# omit the  markers with error rates >5%.
bad_markers <- find_markerpos(gm.no.bad, names(errors_mar[errors_mar > 5]))
save(bad_markers, file = "data/Jackson_Lab_12_batches/bad_markers.RData")
#drop bad markers
gm_after_qc <- drop_markers(gm.no.bad, names(errors_mar)[errors_mar > 5])

gm_after_qc
save(gm_after_qc, file = paste0("data/Jackson_Lab_12_batches/gm_DO", length(ind_ids(gm_after_qc)) ,"_qc.RData"))
save(e,g,snpg, file = "data/Jackson_Lab_12_batches/e_g_snpg_qc.RData")

#replace id
new.id <- str_split_fixed(ind_ids(gm_after_qc), "_",7)[,6]
names(new.id) <- ind_ids(gm_after_qc)
gm_after_qc <- replace_ids(gm_after_qc, new.id)
gm_after_qc

save(gm_after_qc, file = paste0("data/Jackson_Lab_12_batches/gm_DO", length(ind_ids(gm_after_qc)) ,"_qc_newid.RData"))
```
