==> analysis/13_plot_69k_conditional_m2_qtlmapping.Rmd <==
---
title: "plot_69k_conditional_m2.Novelty_residuals_RankNormal_datarelease_12182019"
author: "Hao He"
date: "`r Sys.Date()`"
output: html_document
---

# plot scan result for conditional.69k.m2.Novelty_residuals_RankNormal_datarelease_12182019
```{r plot scan result for conditional.69k.m2.Novelty_residuals_RankNormal_datarelease_12182019, fig.height=14, fig.width=14, eval=TRUE}
#   This script performs plotting on qtl 69k conditional mapping for DO mice m2, Novelty_residuals_RankNormal_datarelease_12182019
library(qtl2)
library(ggplot2)
load("data/Jackson_Lab_11_batches/gm_DO2816_qc.RData")
load("data/69k_grid_pgmap.RData")
pheno <- read.csv("data/pheno/Novelty_residuals_RankNormal_datarelease_12182019.csv", header = TRUE)

print("m2")
load("output/conditional.m2.69k.Novelty_residuals_RankNormal_datarelease_12182019.RData")

#permutation results
#The default is to return the 5% significance thresholds. Thresholds for other (or for multiple) significance levels can be obtained via the alpha argument.
operm <- list()
for(i in c(1:50)[c(-4,-8,-14,-15,-18,-28,-31)]){
  operm[[i]] <- get(load(paste0("output/permu/m2_69k_Novelty_residuals_RankNormal_datarelease_12182019_", colnames(pheno)[14], "_", i, ".RData")))
  operm[[i]] <- operm[[i]]$OFA.total.distance.traveled
}
operm <- matrix(do.call(rbind, operm[c(1:50)[c(-4,-8,-14,-15,-18,-28,-31)]]),ncol = 1)
class(operm) <- c("scan1perm", "matrix")
cutoff <- summary(operm, c(0.05, 0.1))[,1]

#genome-wide results
pdf(file = paste0("output/conditional.m2.69k.Novelty_residuals_RankNormal_datarelease_12182019", "_genomescan.pdf"), width = 14)
for(i in names(condi.m2.qtl.out)){
  par(mar=c(5.1, 4.1, 1.1, 1.1))
  ymx <- maxlod(condi.m2.qtl.out[[i]])*1.2 # overall maximum LOD score
  ymax <- ifelse(ymx > cutoff[1]*1.2, ymx, cutoff[1]*1.2)
  plot(condi.m2.qtl.out[[i]], gmap, lodcolumn=1,
       col="slateblue", ylim=c(0, ymax),
       main=i)
  abline(h=cutoff[1], col="red",lty=2) # 0.05 cutoff
  abline(h=cutoff[2], col="red",lty=1) # 0.1 cutoff
}
dev.off()

#load blup results
x <- list()
#colnames(pheno)[14:41]
for(i in colnames(pheno)[14:41]){
  y <- get(load(paste0("output/blup/conditional.m2.69k.blup_Novelty_residuals_RankNormal_datarelease_12182019_",i,".RData")))
  x[[i]] <- y[[i]]
}
condi.m2.qtl.blup = x

#coeffects plot
condi.m2.qtl.peak <- list()
pdf(file = paste0("output/conditional.m2.69k.Novelty_residuals_RankNormal_datarelease_12182019", "_coeffgeneplot.pdf"), width = 14)
#names(condi.m2.qtl.out)
for(i in names(condi.m2.qtl.out)){
  #peak
  condi.m2.qtl.peak[[i]] <- find_peaks(condi.m2.qtl.out[[i]], gmap,
                                       threshold=maxlod(condi.m2.qtl.out[[i]])-0.001,
                                       drop = 1.5)
  condi.m2.qtl.peak[[i]]$lodcolumn <- i
  print(condi.m2.qtl.peak[[i]])
  
  chr <- as.character(condi.m2.qtl.peak[[i]][condi.m2.qtl.peak[[i]]$lodcolumn == i,"chr"])
  par(mar=c(4.1, 4.1, 1.6, 1.6))
#  plot_coefCC(condi.m2.coef[[i]],
#               gmap[chr],
#               scan1_output=condi.m2.qtl.out[[i]],
#               bgcolor="gray95", #legend="bottomleft",
#               main =i)
  plot_coefCC(condi.m2.qtl.blup[[i]],
              gmap[chr],
              scan1_output=condi.m2.qtl.out[[i]],
              bgcolor="gray95", #legend="bottomleft",
              main =i)
#   plot(condi.m2.snps[[i]]$lod,
#         condi.m2.snps[[i]]$snpinfo,
#         drop_hilit=1.5,
#         genes=condi.m2.genes[[i]],
#         main = i)
}
dev.off()

for(i in names(condi.m2.qtl.out)){
  print(i)
  if(is.null(condi.m2.snps[[i]]))
   next
  top_1 <- data.frame(pheno = i)
  top <- top_snps(condi.m2.snps[[i]]$lod, condi.m2.snps[[i]]$snpinfo)
  top <- cbind(top_1, top)
  write.table(top, file = "output/conditional.69k.Novelty_residuals_RankNormal_datarelease_12182019-variantlist.csv", sep = ",",
              row.names = F, col.names = T, append = TRUE)
  
  topgene <- cbind(top_1, condi.m2.genes[[i]])
  write.table(topgene, file = "output/conditional.69k.Novelty_residuals_RankNormal_datarelease_12182019-genelist.csv", sep = ",",
              row.names = F, col.names = T, append = TRUE)
}
```
